+++ lib/src/result.dart
@@ -1,104 +1,59 @@
-import 'package:equatable/equatable.dart';
+import 'package:freezed_annotation/freezed_annotation.dart';
+part 'result.freezed.dart';
 
-typedef FlatMapFunction<A, B> = int Function(double a);
-
-/// A simple Result wrapper.
-///
-/// You can declare Success and Failure Types as Generic Type..
-class Result<Success, Failure> extends Equatable {
-  final bool _isSuccess;
-  final Success? _value;
-  final Failure? _failure;
-
-  const Result._({
-    required bool isSuccess,
-    Success? value,
-    Failure? failure,
-  })  : _isSuccess = isSuccess,
-        _value = value,
-        _failure = failure;
-
-  /// Marks this Result as successful.
-  factory Result.success(Success value) {
-    return Result._(isSuccess: true, value: value);
+@freezed
+class Result<Success, Failure> with _$Result<Success, Failure> {
+  const Result._();
+  const factory Result.success(Success success) = _Success;
+  const factory Result.failure(Failure failure) = _Failure;
+  bool get isSuccess {
+    return when(success: (_) => true, failure: (_) => false);
   }
 
-  /// Marks this Result as failing
-  factory Result.failure(Failure failure) {
-    return Result._(isSuccess: false, failure: failure);
+  bool get isFailure {
+    return when(success: (_) => false, failure: (_) => true);
   }
 
-  /// returns true, if the result is in success state.
-  bool get isSuccess => _isSuccess;
+  Success? get success =>
+      when(success: (success) => success, failure: (failure) => null);
+  Failure? get failure =>
+      when(success: (success) => null, failure: (failure) => failure);
 
-  /// returns true, if the result is in failure state.
-  bool get isFailure => _isSuccess == false;
-
-  /// returns the success value or null if there is a failure.
-  Success? get success => _value;
-
-  /// returns the failure or null if the state is success.
-  Failure? get failure => _failure;
-
-  /// Somehow like a switch case statement. You can return a Value from the
-  /// functions success and failure.
-  ResultType when<ResultType>({
-    required ResultType Function(Success?) success,
-    required ResultType Function(Failure?) failure,
-  }) {
-    if (_isSuccess) {
-      return success(_value);
-    }
-
-    return failure(_failure);
-  }
-
-  Success? getOrElse<ResultType>(Success Function() successFunction) {
-    if (_isSuccess) {
-      return _value;
-    }
-    return successFunction();
-  }
-
-  /// Use map to convert a Result from one value type to another.
-  /// Changes only value type, not the Failure.
-  Result<ResultType, Failure?> map<ResultType>(
-    ResultType Function(Success?) mapper,
-  ) {
-    if (_isSuccess) {
-      return Result<ResultType, Failure>.success(mapper(_value));
-    } else {
-      return Result<ResultType, Failure?>.failure(_failure);
-    }
-  }
-
-  // Either<L, R2> flatMap<R2>(Function1<R, Either<L, R2>> f) => fold(left, f);
-  /*Result<ResultType, Failure> flatMap<ResultType>(
-      Function1<ResultType, Failure> f) {
-    if (_isSuccess) {}
-  }*/
   Result<ResultType, Failure?> flatMap<ResultType>(
     Result<ResultType, Failure> Function(Success?) mapper,
   ) {
-    if (_isSuccess) {
-      final result = mapper(_value);
-      return result;
-    }
-    return Result<ResultType, Failure?>.failure(_failure);
+    return when(
+      success: (success) => mapper(success),
+      failure: (failure) => Result.failure(failure),
+    );
+  }
+
+  Success? getOrElse<ResultType>(Success Function() successFunction) {
+    return when(
+      success: (success) => success,
+      failure: (failure) => successFunction(),
+    );
+  }
+
+  Result<ResultType, Failure?> mapSuccess<ResultType>(
+    ResultType Function(Success?) mapper,
+  ) {
+    return when(
+      success: (success) =>
+          Result<ResultType, Failure>.success(mapper(success)),
+      failure: (failure) => Result<ResultType, Failure?>.failure(failure),
+    );
   }
 
   @override
-  List<Object?> get props => [_value, _failure, _isSuccess];
-
-  @override
   String toString() {
-    if (_isSuccess) {
-      return _value == null
+    return when(
+      success: (success) => success == null
           ? "Success: (null)"
-          : 'Success: ${_value.toString()}';
-    }
-    return _failure == null
-        ? 'Failure: (null)'
-        : 'Failure: ${_failure.toString()}';
+          : 'Success: ${success.toString()}',
+      failure: (failure) => failure == null
+          ? 'Failure: (null)'
+          : 'Failure: ${failure.toString()}',
+    );
   }
 }